+++ packages/webview_flutter/webview_flutter_android/pigeons/android_webview.dart
@@ -54,6 +54,12 @@ abstract class WebViewHostApi {
     Map<String, String> headers,
   );
 
+  void postUrl(
+    int instanceId,
+    String url,
+    Uint8List data,
+  );
+
   String getUrl(int instanceId);
 
   bool canGoBack(int instanceId);